requirements.txt
streamlit
google-generative-ai
fpdf
sqlite3
bcrypt